+++ src/gui/list_search/controls.rs
@@ -2,6 +2,7 @@ use crate::command::hoard_command::{HoardCommand, Parameterized};
 use crate::gui::commands_gui::{ControlState, DrawState, State};
 use termion::event::Key;
 
+#[allow(clippy::too_many_lines)]
 pub fn key_handler(
     input: Key,
     state: &mut State,
@@ -71,6 +72,24 @@ pub fn key_handler(
                 }
             }
             None
+        }
+        Key::Ctrl('x') => {
+            if state.commands.is_empty() {
+                return None;
+            }
+            let selected_command = state
+                .commands
+                .clone()
+                .get(
+                    state
+                        .command_list_state
+                        .selected()
+                        .expect("there is always a selected command"),
+                )
+                .expect("exists")
+                .clone();
+            state.should_delete = true;
+            Some(selected_command)
         }
         // Select command
         Key::Char('\n') => {
@@ -212,6 +231,7 @@ mod test_controls {
             command_list_state: ListState::default(),
             namespace_tab_state: ListState::default(),
             should_exit: false,
+            should_delete: false,
             draw_state: DrawState::Search,
             control_state: ControlState::Search,
             edit_selection: crate::gui::commands_gui::EditSelection::Command,