+++ CHANGES.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.3
 - ✨ Inline command editing in the GUI. Press `<TAB>` to get started. Only editing the command, its description and the tags are supported for now
+- ✨ Inline command deletion in the GUI. Press `<Ctrl-X>` to delete a command in the GUI view
 ## 1.2
 - ✨ You can now synchronize your commands across multiple terminals! Run `hoard sync --help` to start
 ## 1.1.1